Refactor library to minimize memory usage (#1)
Before this change, each connected component would save their own copy of the connected state and each dispatch would run an action through as many reducers as components were connected.

After this refactoring, the states are stored within a global object in the library, once. Instead of dispatching an action to all component reducers, we use a simple useState Hook from React to save a revision number, that is bumped, each time our global state has changed because of a dispatched action. That way the component gets notified to re-render.

We do not bump the revision (and thus re-render the component) if the state didn't change after being sent through the reducer.